+++ frontend/src/pages/FixedMeetingTicketPage/FixedMeetingTicketPage.styles.ts
@@ -1,0 +1,8 @@
+import { css } from '@emotion/react';
+
+export const s_container = css`
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+  width: 100%;
+`;
